taskmgr.s: taskmgr_switch_tasks: set TSS.ESP0 to stack max top
Task switching made ESP0 eventually pass the task's kernel stack bottom.
The delta was caused by saving such value of ESP in
taskmgr_switch_tasks(), that it was always below its initial value upon
interrupt handler entry.

The solution is to save in ESP0 a constant address.  It must be constant
to bear the meaning of stack top upon entry *and* exit from the
interrupt handler, because everything pushed onto the stack is popped
afterwards (on exit).

This solution is in conflict with the current implementation of initial
task entry.  The initial task is regarded as a continuation of the
kernel code execution after taskmgr initialization, and therefore it
must not corrupt the stack pointer, which the current implementation
tries to achieve by saving the current stack top in TSS.ESP0.

To resolve this conflict the initial task must have its own stack.
Starting the scheduler then pops that stack and causes the task to start
running.  The previous stack (in entry.s) is abandoned.  The function
that starts the scheduler must not return.  More specifically, the
initial task entry must not return.