mutex: fix waiting tasks list race condition
The race condition was: one task releases the lock, checks the waiting
tasks list, sees that it is empty; while another task is just about to
add itself to the waiting tasks list. This led to a situation where a
mutex was free, but some task was blocked waiting on it.